requirements.txt
dj-database-url==2.3.0
dj-rest-auth[with_social]==6.0.0
Django==5.1.2
django-allauth==0.61.1
django-anymail==12.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
python-decouple==3.8
sentry-sdk==2.17.0
stripe==11.2.0
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
redis==5.2.0
django-redis==5.4.0
psycopg2-binary==2.9.10
